move input dataset selection into separate file
